+++ axum-extra/src/extract/optional_path.rs
@@ -1,5 +1,5 @@
 use axum::{
-    extract::{path::ErrorKind, rejection::PathRejection, FromRequestParts, Path},
+    extract::{rejection::PathRejection, FromRequestParts, Path},
     RequestPartsExt,
 };
 use serde::de::DeserializeOwned;
@@ -31,9 +31,11 @@ use serde::de::DeserializeOwned;
 ///     .route("/blog/{page}", get(render_blog));
 /// # let app: Router = app;
 /// ```
+#[deprecated = "Use Option<Path<_>> instead"]
 #[derive(Debug)]
 pub struct OptionalPath<T>(pub Option<T>);
 
+#[allow(deprecated)]
 impl<T, S> FromRequestParts<S> for OptionalPath<T>
 where
     T: DeserializeOwned + Send + 'static,
@@ -45,19 +47,15 @@ where
         parts: &mut http::request::Parts,
         _: &S,
     ) -> Result<Self, Self::Rejection> {
-        match parts.extract::<Path<T>>().await {
-            Ok(Path(params)) => Ok(Self(Some(params))),
-            Err(PathRejection::FailedToDeserializePathParams(e))
-                if matches!(e.kind(), ErrorKind::WrongNumberOfParameters { got: 0, .. }) =>
-            {
-                Ok(Self(None))
-            }
-            Err(e) => Err(e),
-        }
+        parts
+            .extract::<Option<Path<T>>>()
+            .await
+            .map(|opt| Self(opt.map(|Path(x)| x)))
     }
 }
 
 #[cfg(test)]
+#[allow(deprecated)]
 mod tests {
     use std::num::NonZeroU32;